cleanup shared azure hosting
-make namespace usage consistent
-split a file
-use constructor injection
-shor more of message mapping
