+++ samples/azure/shared-host/Version_5/HostWorker/EndpointConfig.cs
@@ -1,16 +1,12 @@
-﻿namespace HostWorker
+﻿using NServiceBus;
+
+#region AzureSharedHosting_HostConfiguration
+
+public class EndpointConfig : IConfigureThisEndpoint, AsA_Host
 {
-    using NServiceBus;
-
-    #region AzureSharedHosting_HostConfiguration
-
-    public class EndpointConfig : IConfigureThisEndpoint, AsA_Host
+    public void Customize(BusConfiguration configuration)
     {
-        public void Customize(BusConfiguration configuration)
-        {
-        }
     }
-
-    #endregion
-
 }
+
+#endregion
