+++ samples/azure/shared-host/Version_5/HostWorker/WorkerRole.cs
@@ -1,34 +1,31 @@
-namespace HostWorker
+using System.Diagnostics;
+using Microsoft.WindowsAzure.ServiceRuntime;
+using NServiceBus.Hosting.Azure;
+
+#region AzureSharedHosting_RoleEntryPoint
+
+public class WorkerRole : RoleEntryPoint
 {
-    using System.Diagnostics;
-    using Microsoft.WindowsAzure.ServiceRuntime;
-    using NServiceBus.Hosting.Azure;
+    NServiceBusRoleEntrypoint nsb = new NServiceBusRoleEntrypoint();
 
-    #region AzureSharedHosting_RoleEntryPoint
-
-    public class WorkerRole : RoleEntryPoint
+    public override bool OnStart()
     {
-        NServiceBusRoleEntrypoint nsb = new NServiceBusRoleEntrypoint();
+        nsb.Start();
 
-        public override bool OnStart()
-        {
-            nsb.Start();
+        Trace.TraceInformation("HostWorker has been started");
 
-            Trace.TraceInformation("HostWorker has been started");
-
-            return base.OnStart();
-        }
-
-        public override void OnStop()
-        {
-            Trace.TraceInformation("HostWorker is stopping");
-
-            nsb.Stop();
-            base.OnStop();
-
-            Trace.TraceInformation("HostWorker has stopped");
-        }
+        return base.OnStart();
     }
 
-    #endregion
+    public override void OnStop()
+    {
+        Trace.TraceInformation("HostWorker is stopping");
+
+        nsb.Stop();
+        base.OnStop();
+
+        Trace.TraceInformation("HostWorker has stopped");
+    }
 }
+
+#endregion
