+++ samples/azure/shared-host/Version_5/Receiver/EndpointConfig.cs
@@ -1,20 +1,16 @@
-﻿namespace Receiver
+﻿using NServiceBus;
+using NServiceBus.Logging;
+
+public class EndpointConfig : IConfigureThisEndpoint
 {
-    using NServiceBus;
-    using NServiceBus.Logging;
-    using Shared;
-
-    public class EndpointConfig : IConfigureThisEndpoint
+    public void Customize(BusConfiguration configuration)
     {
-        public void Customize(BusConfiguration configuration)
-        {
-            configuration.UsePersistence<AzureStoragePersistence>();
-            configuration.UseTransport<AzureStorageQueueTransport>()
-                .ConnectionString("UseDevelopmentStorage=true");
+        configuration.UsePersistence<AzureStoragePersistence>();
+        configuration.UseTransport<AzureStorageQueueTransport>()
+            .ConnectionString("UseDevelopmentStorage=true");
 
-            LogManager.Use<DefaultFactory>().Level(LogLevel.Debug);
+        LogManager.Use<DefaultFactory>().Level(LogLevel.Debug);
 
-            NServiceBusFeatures.DisableNotUsedFeatures(configuration);
-        }
+        configuration.DisableNotUsedFeatures();
     }
 }
