+++ samples/azure/shared-host/Version_5/Sender/ConfigureMessageEndpointMappings.cs
@@ -1,26 +1,23 @@
-﻿namespace Sender
+﻿using NServiceBus.Config;
+using NServiceBus.Config.ConfigurationSource;
+
+#region AzureMultiHost_MessageMapping
+public class ConfigureMessageEndpointMappings : IProvideConfiguration<UnicastBusConfig>
 {
-    using NServiceBus.Config;
-    using NServiceBus.Config.ConfigurationSource;
-
-    public class ConfigureMessageEndpointMappings : IProvideConfiguration<UnicastBusConfig>
+    public UnicastBusConfig GetConfiguration()
     {
-        public UnicastBusConfig GetConfiguration()
+        return new UnicastBusConfig
         {
-            return new UnicastBusConfig
+            MessageEndpointMappings = new MessageEndpointMappingCollection
             {
-                #region AzureMultiHost_MessageMapping
-                MessageEndpointMappings = new MessageEndpointMappingCollection
+                new MessageEndpointMapping
                 {
-                    new MessageEndpointMapping
-                    {
-                        Endpoint = "Receiver",
-                        Messages = "Ping, Shared"
-                    }
+                    Endpoint = "Receiver",
+                    Messages = "Ping, Shared"
                 }
+            }
 
-                #endregion
-            };
-        }
+        };
     }
 }
+#endregion
