+++ samples/azure/shared-host/Version_5/Shared/LogEntry.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using Microsoft.WindowsAzure.Storage.Table;
+
+class LogEntry : TableEntity
+{
+    public LogEntry(string endpoint, string message)
+    {
+        PartitionKey = endpoint;
+        RowKey = DateTime.Now.ToString();
+        Message = message;
+    }
+
+    public string Message { get; set; }
+}
