+++ samples/azure/shared-host/Version_5/Shared/NServiceBusFeatures.cs
@@ -1,15 +1,12 @@
-﻿namespace Shared
-{
-    using NServiceBus;
-    using NServiceBus.Features;
+﻿using NServiceBus;
+using NServiceBus.Features;
 
-    public static class NServiceBusFeatures
+public static class NServiceBusFeatures
+{
+    public static void DisableNotUsedFeatures(this BusConfiguration configuration)
     {
-        public static void DisableNotUsedFeatures(this BusConfiguration configuration)
-        {
-            configuration.DisableFeature<Sagas>();
-            configuration.DisableFeature<SecondLevelRetries>();
-            configuration.DisableFeature<TimeoutManager>();
-        }
+        configuration.DisableFeature<Sagas>();
+        configuration.DisableFeature<SecondLevelRetries>();
+        configuration.DisableFeature<TimeoutManager>();
     }
 }
